Adds basic unit tests for authentication
This unit test just tests the middleware in its current state. Ideally
we should only add this functionality to production environment or
if the user specifically needs to have basic auth enabled.

The next bit of work will be to move the logic out of the server.js into
the middleware itself.